Created a control vs acquisition object
This structure divies up acquistiion and control. Now, a single file collects inputs, it has to have the right devices and things, but it is relatively simple, it already has leg position, and all the electrophysiology things necessary. Then the control is incrediby simple and jsut creates empty mat files, to be filled in after the experiment is over.